+++ original-smash/.gitignore
