+++ original-smash/config/default.json
