+++ original-smash/index.js
