+++ original-smash/package.json
@@ -2,10 +2,9 @@
   "name": "smashcrowd",
   "version": "1.0.0",
   "description": "A drafting tool for Smash Bros.",
-  "main": "vue-server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon .",
+    "start": "gulp",
     "build": "gulp sass"
   },
   "repository": {
@@ -42,7 +41,6 @@
     "gulp": "^4.0.0",
     "gulp-load-plugins": "^1.1.0",
     "gulp-postcss": "^8.0.0",
-    "gulp-sass": "^4.0.1",
-    "nodemon": "^1.18.10"
+    "gulp-sass": "^4.0.1"
   }
 }
