+++ original-smash/public/js/app.js
