+++ original-smash/public/js/libs/jquery.js
