+++ original-smash/public/js/libs/what-input.js
