+++ original-smash/scss/_board.scss
