+++ original-smash/scss/_settings.scss
