+++ original-smash/scss/app.scss
