+++ original-smash/src/smashdown-characterfactory.js
