+++ original-smash/src/smashdown-playerfactory.js
