+++ vue-smash/lib/chars.json
